Fixed tipositem, and changed endpoint to api/tipositem/
